decoder: Write avg strength more often to avoid race condition
If recent_strengths and avg_strengths don't have the same list of
ARFCNs, bad things can happen. This can occur if the DB has history for
only one of these two tables. To avoid this, we write both updates to
the db on every single process() call. We don't currently do this update
atomically though, so there's still a risk of out-of-sync occurring if
one write succeeds and the other doesn't.