Expire RSSIs
Right now, we always report every RSSI we've ever seen in our
recent_strengths and rssi(). This isn't good, since that means we always
assume that our old information is "up to date" in the controller. Now,
we only report for what we have most recently heard as our recent
ARFCNs. This removes some history from the decoder, but the controller
can choose to maintain whatever information it wants if this turns out
to be useful/necessary later. I don't think we need to keep track of the
historical maximum for each channel at the moment.