Controller can now have multiple BTS instances.
Introduced a new "BTS" class that encapsulates all the BTS-specific
operations we deal with. A decoder is tied to a BTS instance, so we can
monitor the GSMTAP stream for each BTS and maintain state based on that.
Further, the controller doesn't deal with decoders, just BTS.
Unfortunately, to maintain the global lock on the gsmws db, we still
require controllers to create and initialize each decoder, but we don't
have to deal with it after initialization.