Make BTS represent a single BTS, and use libopenbts where possible
This is a pretty dramatic refactor of the internals of BTS. We now use
libopenbts wherever possible. We also remove all GSMWS logic from the
BTS element itself: this just presents an interface for manipulating the
BTS and providing access to various characteristics about it. A web
server for example could import this to expose a remote API to a BTS.